requirements.txt
Django>=1.10,<1.11
psycopg2>=2.6,<2.7
python-dateutil>=2.5,<2.6
